backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
tensorflow>=2.14.0
numpy>=1.24.3
pillow>=10.0.1
python-multipart>=0.0.6
pandas>=2.0.3
openpyxl>=3.1.2
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-dotenv>=1.0.0
aiofiles>=23.2.1